zmiany zalecone przez sonarqube
