+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/controller/EventController.java
@@ -1,36 +1,31 @@
 package pl.uniwersytetkaliski.studenteventsplatform.controller;
 
-    import jakarta.validation.Valid;
-    import org.springframework.beans.factory.annotation.Autowired;
-    import org.springframework.format.annotation.DateTimeFormat;
-    import org.springframework.http.HttpStatus;
-    import org.springframework.http.ResponseEntity;
-    import org.springframework.security.access.AccessDeniedException;
-    import org.springframework.security.access.prepost.PreAuthorize;
-    import org.springframework.security.core.Authentication;
-    import org.springframework.security.core.context.SecurityContextHolder;
-    import org.springframework.security.core.userdetails.UsernameNotFoundException;
-    import org.springframework.validation.BindingResult;
-    import org.springframework.validation.FieldError;
-    import org.springframework.web.bind.annotation.*;
-    import pl.uniwersytetkaliski.studenteventsplatform.dto.UserDTO;
-    import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDTO.EventCreateDTO;
-    import pl.uniwersytetkaliski.studenteventsplatform.dto.MessageDTO;
-    import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDTO.EventResponseDTO;
-    import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDTO.EventUpdateDTO;
-    import pl.uniwersytetkaliski.studenteventsplatform.model.EventStatus;
-    import pl.uniwersytetkaliski.studenteventsplatform.model.User;
-    import pl.uniwersytetkaliski.studenteventsplatform.service.EventService;
-    import pl.uniwersytetkaliski.studenteventsplatform.service.NotificationService;
-    import pl.uniwersytetkaliski.studenteventsplatform.service.UserEventService;
-    import pl.uniwersytetkaliski.studenteventsplatform.service.UserService;
+import jakarta.validation.Valid;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.annotation.*;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.UserDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDto.EventCreateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.MessageDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDto.EventResponseDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.eventDto.EventUpdateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.model.EventStatus;
+import pl.uniwersytetkaliski.studenteventsplatform.service.EventService;
+import pl.uniwersytetkaliski.studenteventsplatform.service.NotificationService;
+import pl.uniwersytetkaliski.studenteventsplatform.service.UserEventService;
+import pl.uniwersytetkaliski.studenteventsplatform.service.UserService;
 
-    import java.time.LocalDate;
-    import java.util.HashMap;
-    import java.util.List;
-    import java.util.Map;
+import java.time.LocalDate;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
-    import static java.util.stream.Collectors.toMap;
 
 @RestController
     @RequestMapping("/api/events")
