+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/categoryDto/CategoryUpdateDTO.java
@@ -1,9 +1,9 @@
-package pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDTO;
+package pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 
-public class CategoryCreateDTO {
+public class CategoryUpdateDTO {
     @NotBlank
     @Pattern(regexp = "^\\D*$")
     private String name;
