+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/reportDto/ParticipantCountReportDTO.java
@@ -1,4 +1,4 @@
-package pl.uniwersytetkaliski.studenteventsplatform.dto.reportDTO;
+package pl.uniwersytetkaliski.studenteventsplatform.dto.reportDto;
 
 public class ParticipantCountReportDTO {
     private long participantCount;
