+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/mapper/LocationMapper.java
@@ -1,9 +1,9 @@
 package pl.uniwersytetkaliski.studenteventsplatform.mapper;
 
 import org.springframework.stereotype.Component;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDTO.LocationCreateDTO;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDTO.LocationResponseDTO;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDTO.LocationUpdateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDto.LocationCreateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDto.LocationResponseDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.locationDto.LocationUpdateDTO;
 import pl.uniwersytetkaliski.studenteventsplatform.model.Location;
 
 @Component
