+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/model/Comment.java
@@ -1,8 +1,6 @@
 package pl.uniwersytetkaliski.studenteventsplatform.model;
 
 import jakarta.persistence.*;
-
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
