+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/model/Event.java
@@ -1,12 +1,9 @@
 package pl.uniwersytetkaliski.studenteventsplatform.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 @Entity
