+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/repository/CommentRepository.java
@@ -1,7 +1,6 @@
 package pl.uniwersytetkaliski.studenteventsplatform.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import pl.uniwersytetkaliski.studenteventsplatform.model.Comment;
 
 import java.util.List;
