+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/repository/RatingRepository.java
@@ -3,8 +3,6 @@ package pl.uniwersytetkaliski.studenteventsplatform.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
-import pl.uniwersytetkaliski.studenteventsplatform.model.Comment;
 import pl.uniwersytetkaliski.studenteventsplatform.model.Event;
 import pl.uniwersytetkaliski.studenteventsplatform.model.Rating;
 import pl.uniwersytetkaliski.studenteventsplatform.model.User;
