+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/repository/UserRepository.java
@@ -2,7 +2,6 @@ package pl.uniwersytetkaliski.studenteventsplatform.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import pl.uniwersytetkaliski.studenteventsplatform.model.User;
 
 import java.util.Optional;
