+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/service/CategoryService.java
@@ -3,9 +3,9 @@ package pl.uniwersytetkaliski.studenteventsplatform.service;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDTO.CategoryResponseDTO;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDTO.CategoryCreateDTO;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDTO.CategoryUpdateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDto.CategoryResponseDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDto.CategoryCreateDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.categoryDto.CategoryUpdateDTO;
 import pl.uniwersytetkaliski.studenteventsplatform.mapper.CategoryMapper;
 import pl.uniwersytetkaliski.studenteventsplatform.model.Category;
 import pl.uniwersytetkaliski.studenteventsplatform.repository.CategoryRepository;
