+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/service/ReportService.java
@@ -1,22 +1,22 @@
 package pl.uniwersytetkaliski.studenteventsplatform.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import pl.uniwersytetkaliski.studenteventsplatform.dto.reportDTO.ParticipantPerEventDTO;
+import pl.uniwersytetkaliski.studenteventsplatform.dto.reportDto.ParticipantPerEventDTO;
 import pl.uniwersytetkaliski.studenteventsplatform.repository.EventRepository;
 import pl.uniwersytetkaliski.studenteventsplatform.repository.UserEventRepository;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
 public class ReportService implements ReportServiceInterface {
+    private final EventRepository eventRepository;
+    private final UserEventRepository userEventRepository;
 
-    @Autowired
-    private EventRepository eventRepository;
-    @Autowired
-    private UserEventRepository userEventRepository;
+    public ReportService(EventRepository eventRepository, UserEventRepository userEventRepository) {
+        this.eventRepository = eventRepository;
+        this.userEventRepository = userEventRepository;
+    }
 
     @Override
     public long countEventsBetweenAndDeletedFalseAndAcceptedTrue(LocalDateTime fromDate, LocalDateTime toDate) {
